Move java to lang/
